Cleanup on the shared axis motion code
